requirements.txt
beautifulsoup4==4.11.1
boto3==1.24.68
botocore==1.27.68
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.12
